added dirty version of ordering functions for Expression + added ZeroOrOne operation to NfaBuilder
- Expression's constructor now sorts the regular expression correctly for the NfaBuilder. Parentheses and the operation priorities are taken into account.
- I've put everything into the Expression files. That should be separated.
- More tests are needed.

+++ inc/NfaState.h
@@ -20,7 +20,7 @@ namespace regex {
     // std::nullopt: type != ch or character is wildcard character and matches any character
     std::optional<char> ch;
     std::vector<NfaState*> nextStates;
-    int nSteps = 0;
+    size_t nSteps = 0;
   };
 
   struct NfaFragment {
